requirements.txt
alembic==1.0.8
backcall==0.1.0
blinker==1.4
Click==7.0
decorator==4.4.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.9.0
ipython==7.3.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.1
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2==2.7.7
ptyprocess==0.6.0
Pygments==2.3.1
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.1
traitlets==4.3.2
visitor==0.1.3
wcwidth==0.1.7
Werkzeug==0.14.1
WTForms==2.2.1
